Fix exception thrown by 8<echo

+++ src/main/java/me/manossef/scissors/commands/EchoCommand.java
@@ -2,11 +2,20 @@ package me.manossef.scissors.commands;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
 import me.manossef.scissors.ChatCommandSource;
 import me.manossef.scissors.Commands;
 import net.dv8tion.jda.api.entities.Message;
+import net.dv8tion.jda.api.entities.channel.middleman.GuildChannel;
+import net.dv8tion.jda.api.entities.channel.unions.MessageChannelUnion;
+import net.dv8tion.jda.api.exceptions.InsufficientPermissionException;
 
 public class EchoCommand {
+
+    private static final SimpleCommandExceptionType CANNOT_DELETE = new SimpleCommandExceptionType(() -> "Cannot delete this type of message");
+    private static final SimpleCommandExceptionType NOT_IN_GUILD = new SimpleCommandExceptionType(() -> "Cannot delete messages in channels outside servers");
+    private static final SimpleCommandExceptionType NO_PERMISSION = new SimpleCommandExceptionType(() -> "Cannot delete messages in this channel; no permission");
 
     public static void register(CommandDispatcher<ChatCommandSource> dispatcher) {
 
@@ -18,11 +27,22 @@ public class EchoCommand {
 
     }
 
-    private static int echo(ChatCommandSource source, String message) {
+    private static int echo(ChatCommandSource source, String message) throws CommandSyntaxException {
 
         Message commandMessage = source.commandMessage();
-        commandMessage.delete().queue();
-        commandMessage.getChannel().sendMessage(message).queue();
+        if(!commandMessage.getType().canDelete()) throw CANNOT_DELETE.create();
+        MessageChannelUnion channel = commandMessage.getChannel();
+        if(!(channel instanceof GuildChannel)) throw NOT_IN_GUILD.create();
+        try {
+
+            commandMessage.delete().queue();
+
+        } catch(InsufficientPermissionException e) {
+
+            throw NO_PERMISSION.create();
+
+        }
+        channel.sendMessage(message).queue();
         return 1;
 
     }
